refactor: AnnouncementAddImg를 children으로 받게 변경
