+++ service-manager/src/components/announcement/AnnouncementUpdate.tsx
@@ -5,6 +5,7 @@ import { useAnnounceUpdate } from '../../hooks/react-query/useAnnounceForm';
 import { useAnnounceDetailQuery } from '../../hooks/react-query/useAnnounce';
 import ErrorBoundary from '../common/ErrorBoundary';
 import { AnnouncementImgList } from './AnnouncementImg';
+import { AnnouncementAddImg } from './AnnouncementAddImg';
 
 interface AnnouncementUpdateProps {
   announceId: number;
@@ -78,11 +79,9 @@ export const AnnouncementUpdate = ({ announceId }: AnnouncementUpdateProps) => {
       {images && (
         <div className="mt-8 flex flex-col gap-3">
           <Txt size="h4">이미지 수정</Txt>
-          <AnnouncementImgList
-            isEditPage
-            setImages={setImages}
-            images={images}
-          />
+          <AnnouncementImgList isEditPage setImages={setImages} images={images}>
+            <AnnouncementAddImg setImages={setImages} />
+          </AnnouncementImgList>
         </div>
       )}
